+++ apps/client/src/apis/prompt/type.ts
