+++ apps/client/src/apis/user/queries.ts
@@ -1,6 +1,6 @@
+import { api } from '@apis/instance';
 import { useQuery } from '@tanstack/react-query';
 
-import { api } from '@shared/apis/instance';
 import { components } from '@shared/types/schema';
 
 import { USER_END_POINT } from './end-point';
