+++ apps/client/src/pages/landing/landing-page.tsx
@@ -4,7 +4,7 @@ import { Navigate, useNavigate } from 'react-router';
 
 import { Icon } from '@cds/icon';
 
-import { getAccessToken } from '@shared/storage/token-storage';
+import { getAccessToken } from '@shared/utils/token-storage';
 
 import * as styles from './landing-page.css';
