+++ apps/client/src/pages/new-memo/apis/queries.ts
@@ -1,8 +1,7 @@
+import { api } from '@apis/instance';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { ALL_MEMO_KEY } from '@pages/all-memo/apis/query-key';
-
-import { api } from '@shared/apis/instance';
 
 import { MEMO_END_POINT } from './end-point';
 import { MEMO_KEY } from './query-key';
