+++ apps/client/src/pages/new-memo/components/tab/tab.tsx
@@ -1,11 +1,11 @@
 import { CSSProperties } from 'react';
-
-import { Icon } from '@cds/icon';
-
 import {
   LABEL_COLOR_BY_TEXT,
   PRIMARY_COLOR_VALUE_BY_LABEL_COLOR,
-} from '@shared/constants/label-match';
+} from '@constants/label-match';
+
+import { Icon } from '@cds/icon';
+
 import { LabelTextType } from '@shared/types/label-type';
 
 import * as styles from './tab.css';
