+++ apps/client/src/router/routes/public-route.ts
@@ -2,7 +2,7 @@ import { PATH } from '@router/path';
 import { RouteObject } from 'react-router';
 
 import { LandingPage } from '@pages/landing';
-import LoginCallbackPage from '@pages/login/login-callback-page';
+import LoginCallbackPage from '@pages/login-callback/login-callback-page';
 
 import { LoginPage } from '../lazy';
