+++ apps/client/src/shared/components/memo-list-view/components/tree-view/components/tree-memo-list/tree-memo-list.tsx
@@ -1,4 +1,5 @@
-import { LABEL_COLOR_BY_TEXT } from '@shared/constants/label-match';
+import { LABEL_COLOR_BY_TEXT } from '@constants/label-match';
+
 import { LabelTextType } from '@shared/types/label-type';
 import { StructureMemoTypes } from '@shared/types/memo-info-type';
