+++ apps/client/src/shared/components/memo-list-view/components/tree-view/components/tree-memo/apis/queries.ts
@@ -1,6 +1,6 @@
+import { api } from '@apis/instance';
 import { useQuery } from '@tanstack/react-query';
 
-import { api } from '@shared/apis/instance';
 import { LabelTextType } from '@shared/types/label-type';
 
 import { MEMO_MODAL_END_POINT } from './end-point';
