+++ apps/client/tsconfig.json
@@ -5,7 +5,9 @@
     "plugins": [{ "name": "react" }],
     "types": ["vitest/globals", "vite/client"],
     "paths": {
+      "@apis/*": ["src/apis/*"],
       "@app/*": ["src/app/*"],
+      "@constants/*": ["src/constants/*"],
       "@pages/*": ["src/pages/*"],
       "@router/*": ["src/router/*"],
       "@shared/*": ["src/shared/*"]
